added hook useCompanions
