+++ src/components/widgets/Companions/Companions.tsx
@@ -1,60 +1,17 @@
 "use client";
 
-import { FC, useRef, useState } from "react";
-import { gql, useQuery } from "@apollo/client";
+import { FC } from "react";
+
 import { Title } from "@/components/shared/Title";
-import classes from "./Companions.module.css";
 import { Label } from "@/components/shared/Label";
 import { Input } from "@/components/shared/Input";
+import { useCompanions } from "./hooks/useCompanions";
 
-const GET_FIND_USERS = gql`
-    query Query($email: String, $name: String) {
-        findUsers(email: $email, name: $name) {
-            _id
-            name
-            image
-        }
-    }
-`;
+import classes from "./Companions.module.css";
 
 export const Companions: FC = () => {
-    const [firstName, setFirstName] = useState<string>("");
-    const [lastName, setLastName] = useState<string>("");
-    const [email, setEmail] = useState<string>("");
-
-    const fullName = firstName + " " + lastName;
-
-    const { loading, error, data, refetch } = useQuery(GET_FIND_USERS, {
-        variables: {
-            email,
-            name: fullName,
-        },
-    });
-
+    const { handleFirstNameChange, handleLastNameChange, handleEmailChange, data } = useCompanions();
     console.log(data);
-
-    // if (loading) return <p>Loading...</p>;
-    // if (error) return <p>Error : {error.message}</p>;
-
-    const handleFirstNameChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setFirstName(e.target.value);
-        refetch();
-    };
-
-    const handleLastNameChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setLastName(e.target.value);
-        refetch();
-    };
-
-    // const handleEmailChange = (email: string) => {
-    //     setEmail(email);
-    //     // refetch();
-    // };
-
-    const handleEmailChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setEmail(e.target.value);
-    };
-
     return (
         <div className={classes.searchCompanions}>
             <div className={classes.header}>
@@ -76,7 +33,7 @@ export const Companions: FC = () => {
                     <Input onChange={handleEmailChange} />
                 </Label>
             </div>
-            {data && <div>{data?.userByEmail?.name}</div>}
+            {/* {data && <div>{data?.userByEmail?.name}</div>} */}
         </div>
     );
 };
