+++ src/components/widgets/Companions/hooks/useCompanions.ts
@@ -1,0 +1,36 @@
+import { useState } from "react";
+import { useQuery } from "@apollo/client";
+
+import { GET_FIND_USERS } from "@/app/api/graphql/queries/findUsers";
+
+export const useCompanions = () => {
+    const [firstName, setFirstName] = useState<string>("");
+    const [lastName, setLastName] = useState<string>("");
+    const [email, setEmail] = useState<string>("");
+
+    const name = firstName + " " + lastName;
+
+    const { loading, error, data, refetch } = useQuery(GET_FIND_USERS, {
+        variables: {
+            email,
+            name,
+        },
+    });
+
+    const handleFirstNameChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+        setFirstName(e.target.value);
+        refetch();
+    };
+
+    const handleLastNameChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+        setLastName(e.target.value);
+        refetch();
+    };
+
+    const handleEmailChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+        setEmail(e.target.value);
+        refetch();
+    };
+
+    return { handleFirstNameChange, handleLastNameChange, handleEmailChange, data };
+};
